third_party/blink/web_tests/external/wpt/trusted-types/navigate-to-javascript-url-003-expected.txt

This is a testharness.js-based test.
[FAIL] Setting window.location to a javascript: URL with a default policy that throws should report a CSP violation without rethrowing the exception.
  assert_equals: expected 1 but got 0
Harness: the test ran to completion.
